Calculate operational parameters

Calculate operational parameters

calculate_operational_parameters(
  df,
  calc_list = list(Redox_Out = "(Redox_Out1+Redox_Out2)/2", Redox_Diff =
    "Redox_Out - Redox_In", Power_pump = "Up*Ip", Power_cell = "Uz*Iz", Pump_WhPerCbm =
    "Power_pump/(Flux/1000)", Cell_WhPerCbm = "Power_cell/(Flux/1000)"),
  calc_list_name = c("Mean redox potential in tank",
    "Difference (outflow - inflow) of redox potential", "Power demand of pump",
    "Power demand of cell", "Specific energy demand of pump",
    "Specific energy demand of cell"),
  calc_list_unit = c("mV", "mV", "W", "W", "Wh/m3", "Wh/m3"),
  calc_paras = c("Redox_Out1", "Redox_Out2", "Redox_In", "Flux", "Up", "Ip", "Uz",
    "Iz")
)

Arguments

df
a data frame as retrieved by import_data_haridwar()

calc_list
list with calculation operations to be carried out (default: list(Redox_Out = "(Redox_Out1+Redox_Out2)/2", Redox_Diff = "Redox_Out - Redox_In", Power_pump = "Up*Ip", Power_cell = "Uz*Iz", Pump_WhPerCbm = "Power_pump/Flux/1000", Cell_WhPerCbm = "Power_cell/Flux/1000"))

calc_list_name
full names of parameters to be used for plotting for each calculation specified wit 'calc_list'. default: c('Tank water: Mean redox potential", 'Difference (outflow - inflow) of redox potential', 'Power demand of pump', 'Power demand of cell', 'Specific energy demand of pump', Specific energy demand of cell')

calc_list_unit
units of parameters to be used for plotting for each calculation specified wit 'calc_list'. default: c('mV', 'mV', 'Wh', 'Wh', 'Wh/m3', 'Wh/m3')

calc_paras
a vector with parameter codes used for performing calculations defined in 'calc_list' (default: c('Redox_Out1', 'Redox_Out2', 'Redox_In', 'Flux', 'Up', 'Ip', 'Uz', 'Iz'))

Value

dataframe with calculated operational parameters

Examples

if (FALSE) {
haridwar_raw_list <- import_data_haridwar()
myDat <- calculate_operational_parameters(df = haridwar_raw_list)
}