InfluxDB: write to InfluxDB in Loop

wrapper function for write_to_influxdb

Usage

write_to_influxdb_loop(
  tsv_paths,
  paths,
  changed_only = TRUE,
  max_tsv_files = 5,
  batch_size = 5000
)

Arguments

tsv_paths

vector with tsv_paths with files to be imported by write_to_influxdb which relies on read_pentair_data

paths

paths list with elements raw_data_dir and site_code

changed_only

TRUE if only columns with changing data points within time series of provided tsv_paths (limited by parameter max_tsv_files, i.e. changes between different tsv splits are not detected!) should be written to InfluxDB, otherwise FALSE (default: TRUE)

max_tsv_files

maximum number of tsv files to read at once (should be limited due to high RAM demand), default: 5

batch_size

number of data points that are written in one query (default: 5000)

Value

writes imported data to InfluxDB in Loop